Create README.md
Introduction to Sight Sage Project: A Visionary Approach to Eye Health

In a world increasingly shaped by technology, the Sight Sage project emerges as a beacon of innovation at the intersection of healthcare and artificial intelligence. Focused on revolutionizing the early detection of eye disorders, such as Glaucoma, Diabetic Retinopathy, and Cataract, Sight Sage leverages cutting-edge technologies to provide a comprehensive solution.

At the core of the Sight Sage project is a sophisticated software application, meticulously crafted with the latest web development tools. The frontend of the application is built using React.js, offering a seamless and intuitive user experience. Complementing the frontend, the backend is powered by Node.js, ensuring robust and efficient data processing. Data storage, a critical component of the project, is managed seamlessly with MongoDB, allowing for scalable and flexible storage of vital information.

The heart of the Sight Sage project lies in its machine learning model, specifically designed for the analysis of retinal fundus images. A Convolutional Neural Network (CNN) forms the backbone of this model, trained meticulously on a diverse dataset comprising over 4000 retinal fundus images. To enhance the model's predictive capabilities, K-fold cross-validation was rigorously applied during the training process, ensuring robust performance across a variety of scenarios.

Notably, the Sight Sage project goes beyond mere innovation in its machine learning model. A comprehensive comparative analysis was conducted, evaluating various established CNN models, including ResNet50, DenseNet, VGG16, alongside the project's proprietary model. This comparison sheds light on the unique strengths and weaknesses of each model when applied to the same dataset, providing valuable insights into the efficacy of different approaches in detecting eye disorders.

In essence, Sight Sage not only pioneers a state-of-the-art solution for early detection of eye disorders but also contributes to the scientific discourse by offering a nuanced understanding of the performance of different CNN models in this critical domain. The project's commitment to excellence, innovation, and the well-being of individuals positions Sight Sage as a trailblazer in the pursuit of leveraging technology for enhanced healthcare outcomes.